Cleanups pulled from #29
* Adding RGB endpoint for CR2 files
	* TODO: Convert FITS; timelapse; jpgs
* Removing old dataflow items
* Adding build scripts for containers and kuberenets clusters for the
	similar source finder.
* Removing old Cloud SQL connections
* Misc cleanup

+++ gce-find-similar-sources/similar-source-finder.py
@@ -1,11 +1,14 @@
 import os
 import time
+import re
 import concurrent.futures
 from itertools import zip_longest
 from datetime import datetime
+from tqdm import tqdm
 
 import requests
 from google.cloud import pubsub
+from google.cloud import storage
 import pandas as pd
 
 PROJECT_ID = os.getenv('PROJECT_ID', 'panoptes-survey')
@@ -13,6 +16,10 @@ PROJECT_ID = os.getenv('PROJECT_ID', 'panoptes-survey')
 # Storage
 OBSERVATION_BUCKET_NAME = os.getenv('UPLOAD_BUCKET', 'panoptes-observation-psc')
 PICID_BUCKET_NAME = os.getenv('BUCKET_NAME', 'panoptes-picid')
+
+# Storage
+storage_client = storage.Client(project=PROJECT_ID)
+picid_bucket = storage_client.get_bucket(PICID_BUCKET_NAME)
 
 # Pubsub
 PUBSUB_SUB_PATH = os.getenv('SUB_PATH', 'gce-find-similar-sources')
@@ -28,48 +35,53 @@ update_state_url = os.getenv(
 def main():
     log(f"Starting similar source finder on {pubsub_sub_path}")
 
-    try:
-        flow_control = pubsub.types.FlowControl(max_messages=1)
-        future = subscriber_client.subscribe(
-            pubsub_sub_path, callback=msg_callback, flow_control=flow_control)
+    while True:
+        try:
+            response = subscriber_client.pull(pubsub_sub_path,
+                                              max_messages=1,
+                                              return_immediately=True)
+        except Exception as e:
+            log(f'Problem with pulling from subscriber: {e!r}')
 
-        # Keeps main thread from exiting.
-        log(f"Similar source finder subscriber started, entering listen loop")
-        while True:
+        try:
+            # Process first (and only) message
+            message = response.received_messages[0]
+            log(f"Received message, processing: {message.ack_id}")
+
+            # Acknowledge immediately - resend pubsub on error below.
+            subscriber_client.acknowledge(pubsub_sub_path, [message.ack_id])
+
+            # Will block while processing.
+            process_message(message.message)
+            log(f"Message complete: {message.ack_id}")
+        except IndexError:
+            # No messages, sleep before looping again.
             time.sleep(30)
-    except Exception as e:
-        log(f'Problem starting subscriber: {e!r}')
-        future.cancel()
 
 
 def log(msg):
     print(datetime.now().isoformat(), msg)
 
 
-def msg_callback(message):
+def process_message(message):
 
     attributes = message.attributes
 
-    object_id = attributes['object_id']
+    object_id = attributes['objectId']  # Comes from bucket event.
     sequence_id = attributes['sequence_id']
 
-    # Acknowledge immediately - resend pubsub on error below.
-    message.ack()
+    force_new = attributes.get('force_new', False)
+
     log(f'Received sequence_id: {sequence_id} object_id: {object_id}')
 
-    # Acknowledge the message was received - if we error we will resend message.
-    log(f'Acknowledging message for {sequence_id}')
-    message.ack()
-
-    if 'similar-sources.csv' in object_id:
-        return
-
     if sequence_id is None or sequence_id == '':
-        if object_id is not None and object_id > '':
-            path_parts = object_id.split('/')
-            sequence_id = '_'.join(path_parts[0:3]).replace('.csv', '')
+        log(f'No sequence_id found, looking for matching object_id')
+        matches = re.fullmatch('(PAN.{3}[/_].*[/_]20.{6}T.{6}).csv', object_id)
+        if matches is not None:
+            sequence_id = matches.group(1)
+            log(f'Matched object_id, setting sequence_id={sequence_id}')
         else:
-            log(f'Invalid sequence_id and no object_id: {sequence_id}')
+            log(f'Invalid sequence_id and no object_id, exiting: {object_id}')
             return
 
     # Put sequence_id into path form.
@@ -81,27 +93,29 @@ def msg_callback(message):
 
     # Create the observation PSC
     try:
-        # Update state
-        state = 'psc_file_created'
-        log(f'Updating state for {sequence_id} to {state}')
-        requests.post(update_state_url, json={'sequence_id': sequence_id, 'state': state})
-
+        attributes['sequence_id'] = sequence_id
         psc_df = make_observation_psc_df(**attributes)
         if psc_df is None:
             raise Exception(f'Sequence ID: {sequence_id} No PSC created')
+    except FileNotFoundError as e:
+        log(f'File for {sequence_id} not found in bucket, skipping.')
+        return
     except Exception as e:
         log(f'Error making PSC: {e!r}')
+        # Update state
+        state = 'error_filtering_observation_psc'
+        log(f'Updating state for {sequence_id} to {state}')
+        requests.post(update_state_url, json={'sequence_id': sequence_id, 'state': state})
         return
 
     log(f'Sequence ID: {sequence_id} Done creating PSC, finding similar sources.')
 
     try:
-        find_similar_sources(psc_df, sequence_id)
-
-        # Update state
-        state = 'similar_sources_found'
-        log(f'Updating state for {sequence_id} to {state}')
-        requests.post(update_state_url, json={'sequence_id': sequence_id, 'state': state})
+        if find_similar_sources(psc_df, sequence_id, force_new=force_new):
+            # Update state
+            state = 'similar_sources_found'
+            log(f'Updating state for {sequence_id} to {state}')
+            requests.post(update_state_url, json={'sequence_id': sequence_id, 'state': state})
     except Exception as e:
         log(f'ERROR Sequence {sequence_id}: {e!r}')
     finally:
@@ -109,7 +123,7 @@ def msg_callback(message):
         return
 
 
-def make_observation_psc_df(sequence_id, min_num_frames=10, frame_threshold=0.98, **kwargs):
+def make_observation_psc_df(sequence_id=None, min_num_frames=10, frame_threshold=0.95, **kwargs):
     """Makes a PSC dataframe for the given sequence id.
 
     Args:
@@ -120,6 +134,9 @@ def make_observation_psc_df(sequence_id, min_num_frames=10, frame_threshold=0.98
     Returns:
         `pandas.DataFrame`: Dataframe of the PSC.
     """
+    if sequence_id is None:
+        return
+
     log(f'Sequence ID: {sequence_id} Making PSC')
 
     sequence_id = sequence_id.replace('_', '/')
@@ -151,7 +168,7 @@ def make_observation_psc_df(sequence_id, min_num_frames=10, frame_threshold=0.98
     def has_frame_count(grp):
         return grp.count()['pixel_00'] >= min_frame_count
 
-    # Do the actual fileter an reset the index
+    # Do the actual filter and reset the index
     log(f'Sequence: {sequence_id} filtering sources')
     psc_df = psc_df.reset_index() \
         .groupby('picid') \
@@ -167,42 +184,58 @@ def make_observation_psc_df(sequence_id, min_num_frames=10, frame_threshold=0.98
     return psc_df
 
 
-def do_normalize(params):
+def compare_stamps(params):
+    import numpy as np
+
     try:
         target_params = params[0]
         call_params = params[1]
 
         picid = target_params[0]
-        row = target_params[1]
+        target_table = target_params[1]
 
-        norm_df = call_params['all_psc']
+        ref_table = call_params['all_psc']
         sequence_id = call_params['sequence_id']
+        force_new = call_params['force_new'].lower() == 'true'
 
-        norm_target = row.droplevel('picid')
-        norm_group = ((norm_df - norm_target)**2).dropna().sum(axis=1).groupby('picid')
+        # Check if file already exists
+        save_fn = f'{picid}/{sequence_id}-similar-sources.csv'
+        if force_new is False and picid_bucket.get_blob(save_fn).exists():
+            raise FileExistsError('File already found in storage bucket')
 
-        top_matches = norm_group.sum().sort_values()[:200]
+        # Align index with target
+        include_frames = ref_table.index.levels[0].isin(target_table.index.levels[0])
+        # Get PSC for matching frames
+        ref_table = np.array(ref_table.loc[include_frames])
+
+        # norm_target = target_table.droplevel('picid')
+        norm_group = ((ref_table - target_table)**2).dropna().sum(axis=1).groupby('picid')
+
+        top_matches = (norm_group.sum() / norm_group.std()).sort_values()[:200]
 
         save_fn = f'gs://{PICID_BUCKET_NAME}/{picid}/{sequence_id}-similar-sources.csv'
         top_matches.index.name = 'picid'
         top_matches.to_csv(save_fn, header=['sum_ssd'])
+    except FileExistsError:
+        pass
     except Exception as e:
-        print(f'ERROR Sequence {sequence_id}: {e!r}')
-        state = 'error_finding_similar_sources'
-        print(f'Updating state for {sequence_id} to {state}')
-        requests.post(update_state_url, json={'sequence_id': sequence_id, 'state': state})
+        print(f'ERROR Sequence {sequence_id} Normalize: {e!r}')
     finally:
         return picid
 
 
-def find_similar_sources(stamps_df, sequence_id):
+def find_similar_sources(stamps_df, sequence_id, force_new=False):
     log(f'Loading PSC CSV for {sequence_id}')
 
     # Normalize each stamp
     log(f'Normalizing PSC for {sequence_id}')
     norm_df = stamps_df.copy().apply(lambda x: x / stamps_df.sum(axis=1))
 
-    call_params = dict(all_psc=norm_df, sequence_id=sequence_id)
+    call_params = dict(
+        all_psc=norm_df,
+        sequence_id=sequence_id,
+        force_new=force_new
+    )
 
     log(f'Starting loop of PSCs for {sequence_id}')
     # Run everything in parallel.
@@ -211,13 +244,14 @@ def find_similar_sources(stamps_df, sequence_id):
 
         params = zip_longest(grouped_sources, [], fillvalue=call_params)
 
-        try:
-            picids = list(executor.map(do_normalize, params, chunksize=7))
-            log(f'Found similar stars for {len(picids)} sources')
-        except Exception as e:
-            log(f'ERROR Sequence {sequence_id}: {e!r}')
+        picids = list(tqdm(
+            executor.map(compare_stamps, params, chunksize=2),
+            total=len(grouped_sources)
+        ))
+        log(f'Found similar stars for {len(picids)} sources')
 
     log(f'Sequence {sequence_id}: finished PICID loop')
+    return True
 
 
 if __name__ == '__main__':
